done all

+++ controllers/feed.js
@@ -2,21 +2,27 @@ const { validationResult } = require("express-validator");
 const Post = require("../models/post");
 const { v4: uuidv4 } = require("uuid");
 const fileHelper = require("../util/file");
-const post = require("../models/post");
 const User = require("../models/user");
 
 exports.getPosts = async (req, res, next) => {
-  console.log("get in to the fetch post");
-  const curPage = req.query.page || 1;
-  const countPosts = await Post.count();
+  try {
+    const curPage = req.query.page || 1;
+    const countPosts = await Post.count();
 
-  const posts = await Post.find()
-    .skip((curPage - 1) * 2)
-    .limit(2);
-  res.status(200).json({
-    posts: posts,
-    totalPosts: countPosts,
-  });
+    const posts = await Post.find()
+      .skip((curPage - 1) * 2)
+      .limit(2);
+    res.status(200).json({
+      posts: posts,
+      totalPosts: countPosts,
+    });
+  } catch (error) {
+    if (!error.statusCode) {
+      error.message = "Server error";
+      error.statusCode = 500;
+    }
+    next(error);
+  }
 };
 
 exports.createPost = async (req, res, next) => {
@@ -46,7 +52,6 @@ exports.createPost = async (req, res, next) => {
     await post.save();
 
     const user = await User.findOne({ _id: req.userId });
-    console.log(post);
     user.posts.push(post);
     await user.save();
     res.status(201).json({
@@ -81,10 +86,15 @@ exports.getPost = async (req, res) => {
   });
 };
 
-exports.deletePost = async (req, res) => {
+exports.deletePost = async (req, res, next) => {
   try {
     const postId = req.params.postId;
     const post = await Post.findOne({ _id: postId });
+    if (post.creator.toString() !== req.userId) {
+      const error = new Error("Not authorized");
+      error.statusCode = 403;
+      throw error;
+    }
     await fileHelper.deleteFile(post.imageUrl);
     await Post.deleteOne({ _id: postId });
     res.status(200).json({ message: "Post is Deleted" });
@@ -101,7 +111,7 @@ exports.editPost = async (req, res, next) => {
   try {
     const title = req.body.title;
     const content = req.body.content;
-    const name = req.body.creator;
+    const creator = req.body.creator;
     const postId = req.params.postId;
 
     const resultValidate = validationResult(req);
@@ -110,6 +120,13 @@ exports.editPost = async (req, res, next) => {
       error.statusCode = 422;
       throw error;
     }
+    if (creator.toString() !== req.userId) {
+      const error = new Error("Not authorized");
+      error.statusCode = 403;
+      throw error;
+    }
+    const post = await Post.findOne({ _id: postId });
+    await fileHelper.deleteFile(post.imageUrl);
     if (!req.file) {
       const error = new Error("no image found");
       error.statusCode = 422;
@@ -117,12 +134,12 @@ exports.editPost = async (req, res, next) => {
     }
     const imageUrl = req.file.path.replaceAll("\\", "/");
 
-    const post = await Post.findOne({ _id: postId });
     await fileHelper.deleteFile(post.imageUrl);
     post.title = title;
     post.content = content;
-    post.creator = { name: name };
+    post.creator = creator;
     post.imageUrl = imageUrl;
+
     await post.save();
     res.status(201).json({
       message: "Update successfully",
